dl-*.py, win10.ps1: Refactor the code
Previously, all Python scripts relied on win10.ps1 to create the temp
directory in the Downloads folder. This resulted in them showing an
error when they were not started through win10.ps1, or if a temp
directory was not found in the Downloads folder.

Now, those Python scripts can be run both ways, with or without
win10.ps1. The PowerShell script sends an argument to the Python script
that includes the path to the temp directory. If that argument is not
passed or the Python script is started directly, it downloads the file
in the script's directory.